+++ src/app/_actions/get-profile-by-email-action.ts
@@ -1,6 +1,6 @@
 "use server";
 
-import { getProfileByEmail } from "~/data-access/users/get-profile-by-email";
+import { getProfileByEmail } from "~/data-access/profiles/get-profile-by-email";
 
 export async function getProfileByEmailAction(email: string) {
   return await getProfileByEmail(email);
