+++ src/app/create-profile/band/_actions/create-profile.action.ts
@@ -1,11 +1,10 @@
 "use server";
 
 import { revalidatePath } from "next/cache";
-import { z } from "zod";
 import {
   BandProfile,
   createBandProfile,
-} from "~/data-access/users/create-band-profile";
+} from "~/data-access/profiles/create-band-profile";
 
 export async function createBandProfileAction(
   formData: BandProfile,
